Set cinder-manage command output to utf-8
Basically when Python detects that it is printing to a terminal sys.stdout.encoding is set to the encoding of the terminal. So when we pipe it assumes ascii encoding so when we have a translated string that contains unicode or we are doing lazy translation and have Message objects that don't support str() it will fail.

This patch explicitly encodes messages into unicode before printing them.

Related-Bug: #1396686